Remove return from csv command
handle() expects an integer return which is treated
as an exit code. Returning a bool will set the exit code
of the command to `1`, which is a generic "error" exit code.

Returning nothing from the method causes the correct `0` code
to be returned

Previous incorrect behaviour:
```
➜ php artisan csv:generate
Generating CSV at SalaryDates.csv
➜ echo $?
1
```